Made the code for bootleg PS3 controller dynamic
Axes and buttons can by dynamically assigned by the player. The left and
right controls can't be assigned by a PS5 DualSense controller because the
DualSense's directional/arrow buttons are button inputs and not axes
inputs. The code expects the left and right controls to be axes inputs.
I need to make the controls accommodate both input types.